core: stdcm: move java module from api to osrd
