+++ core/src/main/java/fr/sncf/osrd/stdcm/graph/STDCMNode.java
@@ -1,4 +1,4 @@
-package fr.sncf.osrd.api.stdcm.graph;
+package fr.sncf.osrd.stdcm.graph;
 
 import fr.sncf.osrd.infra.api.reservation.DiDetector;
