+++ core/src/main/java/fr/sncf/osrd/stdcm/graph/STDCMUtils.java
@@ -1,4 +1,4 @@
-package fr.sncf.osrd.api.stdcm.graph;
+package fr.sncf.osrd.stdcm.graph;
 
 import com.google.common.primitives.Doubles;
 import fr.sncf.osrd.envelope.Envelope;
