+++ core/src/test/java/fr/sncf/osrd/stdcm/STDCMPathfindingBuilder.java
@@ -2,9 +2,7 @@ package fr.sncf.osrd.stdcm;
 
 import com.google.common.collect.ImmutableMultimap;
 import com.google.common.collect.Multimap;
-import fr.sncf.osrd.api.stdcm.OccupancyBlock;
-import fr.sncf.osrd.api.stdcm.STDCMResult;
-import fr.sncf.osrd.api.stdcm.graph.STDCMPathfinding;
+import fr.sncf.osrd.stdcm.graph.STDCMPathfinding;
 import fr.sncf.osrd.envelope_sim.allowances.utils.AllowanceValue;
 import fr.sncf.osrd.infra.api.signaling.SignalingInfra;
 import fr.sncf.osrd.infra.api.signaling.SignalingRoute;
